// src/utils/enum/selectedMenu.enum.ts
/**
 * Enum que define os itens do menu selecionáveis.
 *
 * @enum {string}
 */
export enum SelectedMenu {
  BEGIN = 'Begin',
  WAITING_VOLUNTEER = 'Waiting_volunteer',
  WAITING_CONFIRMATION = 'Waiting_confirmation',
  CLASSES_SCHEDULED = 'Classes_scheduled',
  TUTOR_BEGIN = 'Tutor_begin',
  TUTOR_SCHEDULED = 'Tutor_scheduled',
  TUTOR_REQUEST = 'Tutor_request',
  TUTOR_WAITING_CONFIRMATION = 'Tutor_waiting_confirmation'
}